AIStudyCompanion/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.42.0",
    "cohere>=5.16.1",
    "email-validator>=2.2.0",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-cloud-texttospeech>=2.27.0",
    "gunicorn>=23.0.0",
    "langchain-cohere>=0.4.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "numpy>=2.3.1",
    "openai>=1.95.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
